# LUCENE-9848 Sort HNSW graph neighbors for construction