--- a/lucene/core/src/java/org/apache/lucene/util/hnsw/NeighborArray.java
+++ b/lucene/core/src/java/org/apache/lucene/util/hnsw/NeighborArray.java
@@ -21,29 +21,61 @@ import org.apache.lucene.util.ArrayUtil;
 
 /**
  * NeighborArray encodes the neighbors of a node and their mutual scores in the HNSW graph as a pair
- * of growable arrays.
+ * of growable arrays. Nodes are arranged in the sorted order of their scores in descending order
+ * (if scoresDescOrder is true), or in the ascending order of their scores (if scoresDescOrder is
+ * false)
  *
  * @lucene.internal
  */
 public class NeighborArray {
-
+  private final boolean scoresDescOrder;
   private int size;
 
   float[] score;
   int[] node;
 
-  public NeighborArray(int maxSize) {
+  public NeighborArray(int maxSize, boolean descOrder) {
     node = new int[maxSize];
     score = new float[maxSize];
+    this.scoresDescOrder = descOrder;
   }
 
+  /**
+   * Add a new node to the NeighborArray. The new node must be worse than all previously stored
+   * nodes.
+   */
   public void add(int newNode, float newScore) {
     if (size == node.length - 1) {
       node = ArrayUtil.grow(node, (size + 1) * 3 / 2);
       score = ArrayUtil.growExact(score, node.length);
     }
+    if (size > 0) {
+      float previousScore = score[size - 1];
+      assert ((scoresDescOrder && (previousScore >= newScore))
+              || (scoresDescOrder == false && (previousScore <= newScore)))
+          : "Nodes are added in the incorrect order!";
+    }
     node[size] = newNode;
     score[size] = newScore;
+    ++size;
+  }
+
+  /** Add a new node to the NeighborArray into a correct sort position according to its score. */
+  public void addAndSort(int newNode, float newScore) {
+    if (size == node.length - 1) {
+      node = ArrayUtil.grow(node, (size + 1) * 3 / 2);
+      score = ArrayUtil.growExact(score, node.length);
+    }
+    int insertionPoint =
+        scoresDescOrder
+            ? descSortFindRightMostInsertionPoint(newScore)
+            : ascSortFindRightMostInsertionPoint(newScore);
+    for (int i = size; i > insertionPoint; i--) {
+      node[i] = node[i - 1];
+      score[i] = score[i - 1];
+    }
+    node[insertionPoint] = newNode;
+    score[insertionPoint] = newScore;
     ++size;
   }
 
@@ -72,8 +104,38 @@ public class NeighborArray {
     size--;
   }
 
+  public void removeIndex(int idx) {
+    for (int i = idx; i < (size - 1); i++) {
+      node[i] = node[i + 1];
+      score[i] = score[i + 1];
+    }
+    size--;
+  }
+
   @Override
   public String toString() {
     return "NeighborArray[" + size + "]";
+  }
+
+  private int ascSortFindRightMostInsertionPoint(float newScore) {
+    int start = 0;
+    int end = size - 1;
+    while (start <= end) {
+      int mid = (start + end) / 2;
+      if (score[mid] > newScore) end = mid - 1;
+      else start = mid + 1;
+    }
+    return start;
+  }
+
+  private int descSortFindRightMostInsertionPoint(float newScore) {
+    int start = 0;
+    int end = size - 1;
+    while (start <= end) {
+      int mid = (start + end) / 2;
+      if (score[mid] < newScore) end = mid - 1;
+      else start = mid + 1;
+    }
+    return start;
   }
 }