@mayya-sharipova 2022-05-02
I've run benchmarks on 2 datasets:
baseline  – main branch
candidate – this PR:

Indexing time was the same or slightly better.
Recall slightly increased: 1-3%
QPS slightly decreased

---
**glove-100-angular M:16 efConstruction:100**


|             | baseline recall | baseline QPS | candidate recall | candidate QPS |
| ----------- | --------------: | -----------: | ---------------: | ------------: |
| n_cands=10  |           0.463 |     3742.506 |            0.476 |      3743.109 |
| n_cands=20  |           0.529 |     3191.042 |            0.541 |      2954.978 |
| n_cands=40  |           0.603 |     2576.537 |            0.617 |      2681.529 |
| n_cands=80  |           0.682 |     1877.982 |            0.696 |      1629.359 |
| n_cands=120 |           0.721 |     1492.808 |            0.737 |      1447.522 |
| n_cands=200 |           0.766 |     1099.154 |            0.784 |      1093.390 |
| n_cands=400 |           0.818 |      672.911 |            0.834 |       667.136 |
| n_cands=600 |           0.845 |      489.292 |            0.859 |       478.109 |
| n_cands=800 |           0.862 |      383.402 |            0.875 |       383.488 |


---
**sift-128-euclidean  M:16 efConstruction:500**

|             | baseline recall | baseline QPS | candidate recall | candidate QPS |
| ----------- | --------------: | -----------: | ---------------: | ------------: |
| n_cands=10  |           0.747 |     3891.531 |            0.753 |      3304.833 |
| n_cands=20  |           0.817 |     3359.364 |            0.819 |      2879.256 |
| n_cands=40  |           0.889 |     2590.605 |            0.887 |      2280.737 |
| n_cands=80  |           0.944 |     1798.558 |            0.941 |      1648.794 |
| n_cands=120 |           0.964 |     1383.721 |            0.963 |      1377.711 |
| n_cands=200 |           0.983 |      973.862 |            0.981 |       917.995 |
| n_cands=400 |           0.994 |      586.816 |            0.993 |       548.358 |
| n_cands=600 |           0.997 |      427.128 |            0.996 |       408.959 |
| n_cands=800 |           0.998 |      341.178 |            0.997 |       316.840 |