@mayya-sharipova 2022-05-04
@msokolov Thanks so much for your feedback. I've addressed it in 160904904c94ffd4d194fc2509124c0e2eb9c44a.

I've also did another set of benchmarking with new changes, this time with higher `M` and `efConstruction` values:

glove-100-angular M:32 efConstruction:500

Happy to report that indexing took just slightly more (3%), while for searches reported slight gains in both recall and QPS:

**Indexing Speed**
baseline: Indexed 1183514 documents in 3697s
candidate Indexed 1183514 documents in 3814s


**Search**

|             | baseline recall | baseline QPS | candidate recall | candidate QPS |
| ----------- | --------------: | -----------: | ---------------: | ------------: |
| n_cands=10  |           0.634 |     2628.879 |            0.659 |      2659.589 |
| n_cands=20  |           0.696 |     2034.672 |            0.719 |      2186.015 |
| n_cands=40  |           0.762 |     1611.403 |            0.784 |      1666.735 |
| n_cands=80  |           0.824 |     1150.786 |            0.843 |      1161.111 |
| n_cands=120 |           0.855 |      878.114 |            0.874 |       890.783 |
| n_cands=200 |           0.892 |      610.891 |            0.907 |       621.173 |
| n_cands=400 |           0.931 |      343.649 |            0.944 |       359.924 |
| n_cands=600 |           0.949 |      239.217 |            0.960 |       256.863 |
| n_cands=800 |           0.961 |      189.523 |            0.970 |       199.867 |